+++ greskell/ChangeLog.md
@@ -8,6 +8,7 @@
 
 * Add `unsafeCastKey` function.
 * [BREAKING CHANGE] Remove `ElementID` type family from `Element` class.
+* [BREAKING CHANGE] Add `elementId` and `elementLabel` functions to `Element` class.
 * [BREAKING CHANGE] Remove `Vertex`, `Edge`, `PropertyMap` and
   `FromGraphSONWithKey` classes.
 * [BREAKING CHANGE] Remove `PropertyMapSingle` and `PropertyMapList`